backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
google-generativeai
python-dotenv
pydantic
pydantic-settings
langchain-text-splitters
pypdf
python-docx
sqlalchemy
psycopg2-binary
alembic
